Refactor how we create Talks
We were duplicating some database logic in two places; move this into
the Talk class itself
