+++ app/src/main/java/org/dharmaseed/android/Talk.java
@@ -2,6 +2,8 @@ package org.dharmaseed.android;
 
 import android.content.Context;
 import android.database.Cursor;
+import android.database.sqlite.SQLiteDatabase;
+import android.util.Log;
 
 import java.io.File;
 
@@ -11,6 +13,8 @@ import java.io.File;
  * @author jakewilson
  */
 public class Talk {
+
+    private static final String LOG_TAG = "Talk";
 
     private String title;
     private String description;
@@ -30,44 +34,92 @@ public class Talk {
 
     private double durationInMinutes;
 
-    public Talk(Cursor cursor, Context context) {
-        this.create(cursor);
-        this.context = context;
-    }
-
     /**
      * Create the object by setting all model values from the cursor
      * @param cursor the result from the DB
      */
-    private void create(Cursor cursor) {
-        setTitle(cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.TITLE)).trim());
+    public static Talk lookup(DBManager dbManager, Context context, int talkID) {
+        Talk talk = null;
+        Cursor cursor = getCursor(dbManager, talkID);
+        if (cursor.moveToFirst()) {
+            talk = new Talk();
+            talk.title = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.TITLE)).trim();
+            talk.description = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.DESCRIPTION)).trim();
 
-        setDescription(cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.DESCRIPTION)).trim());
+            String url = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.AUDIO_URL));
+            talk.audioUrl = "https://www.dharmaseed.org" + url;
 
-        String url = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.AUDIO_URL));
-        setAudioUrl("https://www.dharmaseed.org" + url);
+            String recDate = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.RECORDING_DATE));
+            if(recDate == null) {
+                recDate = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.UPDATE_DATE));
+            }
+            talk.date = recDate;
 
-        String recDate = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.RECORDING_DATE));
-        if(recDate == null) {
-            recDate = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.UPDATE_DATE));
+            talk.teacherId = cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Talk.TEACHER_ID));
+            talk.teacherName = cursor.getString(cursor.getColumnIndexOrThrow("teacher_name")).trim();
+            talk.centerName = cursor.getString(cursor.getColumnIndexOrThrow("center_name")).trim();
+            talk.photoFileName = DBManager.getTeacherPhotoFilename(cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Teacher.ID)));
+            talk.retreatId = cursor.getColumnIndexOrThrow(DBManager.C.Talk.RETREAT_ID);
+            talk.durationInMinutes = cursor.getDouble(cursor.getColumnIndexOrThrow(DBManager.C.Talk.DURATION_IN_MINUTES));
+            talk.path = cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.FILE_PATH));
+
+            talk.id = talkID;
+            talk.context = context;
+        } else {
+            Log.e(LOG_TAG, "Could not look up talk, id=" + talkID);
         }
-        setDate(recDate);
-
-        setTeacherId(cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Talk.TEACHER_ID)));
-
-        setTeacherName(cursor.getString(cursor.getColumnIndexOrThrow("teacher_name")).trim());
-
-        setCenterName(cursor.getString(cursor.getColumnIndexOrThrow("center_name")).trim());
-
-        setPhotoFileName(DBManager.getTeacherPhotoFilename(cursor.getInt(cursor.getColumnIndexOrThrow(DBManager.C.Teacher.ID))));
-
-        setRetreatId(cursor.getColumnIndexOrThrow(DBManager.C.Talk.RETREAT_ID));
-
-        setDurationInMinutes(cursor.getDouble(cursor.getColumnIndexOrThrow(DBManager.C.Talk.DURATION_IN_MINUTES)));
-
-        setPath(cursor.getString(cursor.getColumnIndexOrThrow(DBManager.C.Talk.FILE_PATH)));
 
         cursor.close();
+        return talk;
+
+    }
+
+    public static Cursor getCursor(DBManager dbManager, int talkID) {
+
+        SQLiteDatabase db = dbManager.getReadableDatabase();
+        String query = String.format(
+                "SELECT %s, %s.%s, %s, %s, %s, %s, %s, %s, %s, %s.%s AS teacher_name, %s.%s AS center_name, "
+                        + "%s.%s FROM %s, %s, %s WHERE %s.%s=%s.%s AND %s.%s=%s.%s AND %s.%s=%s",
+                DBManager.C.Talk.TITLE,
+                DBManager.C.Talk.TABLE_NAME,
+                DBManager.C.Talk.DESCRIPTION,
+                DBManager.C.Talk.AUDIO_URL,
+                DBManager.C.Talk.DURATION_IN_MINUTES,
+                DBManager.C.Talk.RECORDING_DATE,
+                DBManager.C.Talk.UPDATE_DATE,
+                DBManager.C.Talk.RETREAT_ID,
+                DBManager.C.Talk.FILE_PATH,
+                DBManager.C.Talk.TEACHER_ID,
+                DBManager.C.Teacher.TABLE_NAME,
+                DBManager.C.Teacher.NAME,
+                DBManager.C.Center.TABLE_NAME,
+                DBManager.C.Center.NAME,
+
+                DBManager.C.Teacher.TABLE_NAME,
+                DBManager.C.Teacher.ID,
+
+                // FROM
+                DBManager.C.Talk.TABLE_NAME,
+                DBManager.C.Teacher.TABLE_NAME,
+                DBManager.C.Center.TABLE_NAME,
+
+                // WHERE
+                DBManager.C.Talk.TABLE_NAME,
+                DBManager.C.Talk.TEACHER_ID,
+                DBManager.C.Teacher.TABLE_NAME,
+                DBManager.C.Teacher.ID,
+
+                DBManager.C.Talk.TABLE_NAME,
+                DBManager.C.Talk.VENUE_ID,
+                DBManager.C.Center.TABLE_NAME,
+                DBManager.C.Center.ID,
+
+                DBManager.C.Talk.TABLE_NAME,
+                DBManager.C.Talk.ID,
+                talkID
+        );
+
+        return db.rawQuery(query, null);
     }
 
     public String getTitle() {
@@ -134,55 +186,7 @@ public class Talk {
         return path;
     }
 
-    private void setTitle(String title) {
-        this.title = title;
-    }
-
-    private void setDescription(String description) {
-        this.description = description;
-    }
-
-    private void setAudioUrl(String audioUrl) {
-        this.audioUrl = audioUrl;
-    }
-
-    private void setDurationInMinutes(double durationInMinutes) {
-        this.durationInMinutes = durationInMinutes;
-    }
-
-    private void setDate(String date) {
-        this.date = date;
-    }
-
-    private void setTeacherName(String teacherName) {
-        this.teacherName = teacherName;
-    }
-
-    private void setCenterName(String centerName) {
-        this.centerName = centerName;
-    }
-
-    private void setPhotoFileName(String photoFileName) {
-        this.photoFileName = photoFileName;
-    }
-
     public void setPath(String path) {
         this.path = path;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    private void setVenueId(int venueId) {
-        this.venueId = venueId;
-    }
-
-    private void setTeacherId(int teacherId) {
-        this.teacherId = teacherId;
-    }
-
-    private void setRetreatId(int retreatId) {
-        this.retreatId = retreatId;
     }
 }
